Add MOV file format

+++ utils.py
@@ -6,6 +6,9 @@ import shutil
 import logging
 import filecmp
 from Tkinter import *
+import os
+from datetime import datetime
+
 
 logger = logging.getLogger('PicTool')
 fh = logging.FileHandler('pictool.log')
@@ -63,7 +66,7 @@ def remove(path):
 def get_time():
     return time.time()
 
-def _handle_files(files, _format=['JPG']):
+def _handle_files(files, _format=['MOV']):
     res = []
     for _f in files:
         _f = _f.split(' ')[-1].strip('\r')
@@ -74,18 +77,23 @@ def _handle_files(files, _format=['JPG']):
             res.append(_f)
     return res
 
-def get_files(path, _format=['JPG']):
-    res = []
+def get_files(path, _format=[]):
+    res = {}
+    for f in _format:
+        res[f] = []
     for root, dirs, files in os.walk(path):
         for _f in files:
 	    _f_end = _f.split('.')[-1].strip()
 	    if _f_end.upper() in _format:
-	        res.append(path_join([root, _f]))
+	        res[_f_end.upper()].append(path_join([root, _f]))
     return res
 
 def get_file_size(path):
-    size = os.path.getsize(path)
-    return size
+    try:
+        size = os.path.getsize(path)
+        return size
+    except:
+        return 0
 
 def format_time(sec):
     return time.strftime('%H:%M:%S', time.gmtime(sec))
@@ -94,13 +102,16 @@ def move(src, target):
     cmd = "move '{0}' '{1}'".format(src, target)
     run_cmd(cmd)
 
-def copy(src, target):
+def copy(src, target, ignore_exist=False):
     try:
-        shutil.copy2(src, target)
+        if not ignore_exist or not path_exists(target):
+            shutil.copy2(src, target)
+            return 'Copy {0} to {1}'.format(
+                src.encode('utf-8'), target.encode('utf-8'))
     #shutil.copystat(src, target)
     except:
         pass
-    return 'Copy {0} to {1}'.format(src.encode('utf-8'), target.encode('utf-8'))
+    return 'Exist: {0}'.format(target.encode('utf-8'))
 
 def is_diff(src, target):
     return filecmp.cmp(src, target)
@@ -122,3 +133,6 @@ def create_folder(path):
     os.makedirs(path)
     return 'Create folder: {0}'.format(path.encode('utf-8'))
     
+def get_modify_date(path):
+    time = os.path.getmtime(path)
+    return datetime.fromtimestamp(int(time)).strftime('%Y:%m:%d')
